Removed cast warning Removed old member variable

+++ Week 1/Framework/SDLFramework/SDLFramework/AStarNode.cpp
@@ -1,6 +1,6 @@
 #include "AStarNode.h"
 
-AStarNode::AStarNode(Node *node, int distance, int distanceToTravel){
+AStarNode::AStarNode(Node *node, double distance, double distanceToTravel){
 	mNode = node;
 	mDistance = distance;
 	mDistanceToTravel = distanceToTravel;
