+++ Week 1/Framework/SDLFramework/SDLFramework/AStarNode.h
@@ -6,10 +6,10 @@ class AStarNode{
 private:
 
 public:
-	AStarNode(Node *node, int distance, int distanceToTravel);
+	AStarNode(Node *node, double distance, double distanceToTravel);
 	~AStarNode();
 
-	int mDistance = 1;
-	int mDistanceToTravel = 0;
+	double mDistance = 1;
+	double mDistanceToTravel = 0;
 	Node *mNode = nullptr;
 };
